removed createrecordingsession and moved its logic to devtools
